Step towards getting CMIS sessions created. The text-context.xml declares the beans that check for the existence of users and claims. The CMIS sessions will then be created based on the code and configuration pulled in from the cmis benchmark load test.
The tests are failing:
Failed tests:
  Incorrect number of event names: [start] expected:<2> but was:<1>
  There should be events in the event queue. . Actual: 0

When running the tests, the debug shows the configuration passed to the
different tests, but also the output of some of the test runs.
The test is failing because the 'start' event, which is handled by the
bean 'event.start' is checking that there are enough users to continue,
but there are not.  In the next commit, we will create a user in the
collection.  In a full test environment, some other load test would have
been run that created and recorded the users.