Claims test increment The test takes control of the creation of the test MongoDB database in order to inject a user.  This causes the user check to pass, which leads to the subsequent failure of the claims count check.  Look for this in the test logs:
LifecycleController: 184] - Started components:
servicesLifeCycleController
       BMClaimsTest: 202] - Showing 1 of each type of event:

Alternatively, run the test from the IDE and put a breakpoint at the line
of code and observe the results of the query to get one of each type of
result.  As the load test grows, we need to keep adjusting the test's
expectations of what is going to appear in the results.